learnt fetching image from url in jetpack

+++ app/src/main/java/com/example/weatherapp/models/Wind.kt
@@ -2,5 +2,5 @@ package com.example.weatherapp.models
 
 data class Wind(
     val deg: Int,
-    val speed: Int
+    val speed: Double
 )
